refactor: member 모듈에서 enumerate와 exception 패키지 경로를 common 패키지로 수정
